[EGD-7141] Fix GUI googletest
Fix invalidly added GUI gtest. Test dummies moved to individual files to
deal with a bug in cmake test discovery for gtest.
